refactor(explorer): collapse ExplorerView + S3Browser into one container
One provider-driven Explorer (src/components/explorer/Explorer.tsx) now renders
both SFTP/SCP and S3, via provider ops + usePaneState. Backend-specific bits are
gated by capability/type (sudo -> SFTP; presign -> S3; drag-out/move/copy by
caps). S3's bucket selection moves to a thin S3Explorer gate that then mounts the
shared container. Providers gain breadcrumbs() for the path bar.

usePaneState scopes each store subscription to the backing provider so a
cross-backend clipboard change doesn't re-render unrelated explorers.

Retire ExplorerView, S3Browser, the dead SftpPage/SftpTabs, and the now-orphaned
SftpSessionPicker/PathBar/transferEventName; rewire ExplorerPage + S3Page; move
the upload test onto Explorer.
