+++ src/components/explorer/index.ts
@@ -1,3 +1,4 @@
 export { ExplorerFileTable } from "./ExplorerFileTable";
 export { ExplorerToolbar } from "./ExplorerToolbar";
 export { ExplorerDropZone } from "./ExplorerDropZone";
+export { Explorer } from "./Explorer";
