+++ src/components/s3/index.ts
@@ -1,3 +1,3 @@
 export { S3Page } from "./S3Page";
-export { S3Browser } from "./S3Browser";
+export { S3Explorer } from "./S3Explorer";
 export { S3ConnectDialog } from "./S3ConnectDialog";
