+++ src/components/sftp/index.ts
@@ -1,7 +1,2 @@
-export { SftpPage } from "./SftpPage";
 export { ExplorerPage } from "./ExplorerPage";
-export { SftpTabs } from "./SftpTabs";
-export { ExplorerView } from "./ExplorerView";
-export { SftpSessionPicker } from "./SftpSessionPicker";
-export { PathBar } from "./PathBar";
 export { ExplorerFileTable, ExplorerToolbar, ExplorerDropZone } from "../explorer";
